Initial work towards a 1.16.4 port
A lot of work still lies ahead and maybe only 20% of a fully feldged 1.16.4 port is done.
This patch only fixed around 50% of the compile errors  (although a few compile errors are
 from the libraries - so it's more around 75%) and compile warnings weren't reduced by much
either (outside from removing unused imports). What also is a bit grieving me is that
origin/1.15.2 is not containing the latest updates, which means that I have to rebase
this patch at some point. It might also be an indicator that I'm not really supporting
the most open source project out here and that a 1.16.4 port might be already along
the way to release silently outside of GitHub. And who know's, maybe the 5 hours I
poured into it were a bit useless (although I'd debate that).

Anyways, this patch will make so the project no longer compiles and even if there
were no compile issues I would be suprised if this were to run.

Also contains a mode change for the gradlew file (100644 -> 100755).

Signed-off-by: Geolykt <admin@geolykt.de>