Removed CoconutTree compile errors
While removing half of the functionallity the tree has.
It's of course something that is fixable and as such will be fixed after it compiles
and we can go into the testing phase of this. To be honest I don't think I am
qualified enough to have made this patch, but who would have done it otherwise?

Signed-off-by: Geolykt <admin@geolykt.de>